+++ tests/Mockery/Generator/StringManipulation/Pass/InstanceMockPassTest.php
@@ -40,5 +40,6 @@ class InstanceMockPassTest extends TestCase
         $code = $pass->apply('class Dave { }', $config);
         $this->assertContains('public function __construct', $code);
         $this->assertContains('protected $_mockery_ignoreVerification', $code);
+        $this->assertContains('this->_mockery_constructorCalled(func_get_args());', $code);
     }
 }